month printing bug fix
I have this api which provides the 1st of islamic month, I give georgean date and it gives islamic date with month. the api gives month number from 1 to 12. I have a problem, when I give it 29-03-2025 it gives me the month of SHAWWAL and on the next day it gives output of  ZUL-QADDAH but I want to print it the same month Shawwal. I am using python.